feat: add permission management system
This commit introduces a new permission management system that allows the bot owner to control who can use the bot and in which groups the bot can be used.

New commands have been added to manage allowed users and groups:
- .adduser: adds a user to the allowed list
- .deluser: removes a user from the allowed list
- .addgroup: adds a group to the allowed list
- .delgroup: removes a group from the allowed list

The permission checking logic has been updated to use the new permission manager.